Moved GraphBuilderBaseWithCliques to graph_builder_base.rs
